Refactored dashboard.js > Dashboard.js
It's much more simple now it just gets a config, resolves a dashboard (modules with data) or gets a single module.
