+++ test/Dashboard.spec.js
@@ -1,0 +1,117 @@
+var requireSubvert = require('require-subvert')(__dirname),
+  dashboardResponse = require('./fixtures/dashboard-response.json'),
+  Module = require('../lib/Module'),
+  Q = require('q');
+
+describe('Dashboard', function () {
+  var Dashboard,
+    stub,
+    deferred;
+
+  var module, moduleDataResponse;
+
+  beforeEach(function () {
+    deferred = Q.defer();
+    stub = sinon.stub().returns(deferred.promise);
+    requireSubvert.subvert('../lib/request-promise', stub);
+    Dashboard = requireSubvert.require('../lib/Dashboard');
+    Module.prototype.resolve = stub;
+    module = {
+      'title': 'test',
+      'format': {
+        'type': 'number'
+      },
+      'module-type': 'kpi',
+      'value-attribute': 'specific_data',
+      'data-source': {
+        'data-group': 'transactional-services',
+        'data-type': 'summaries',
+        'query-params': {
+          'sort_by': '_timestamp:descending',
+          'filter_by': [
+            'service_id:bis-accounts-filing',
+            'type:seasonally-adjusted'
+          ]
+        }
+      }
+    };
+
+    moduleDataResponse = [
+      {
+        '_quarter_start_at': '2013-07-01T00:00:00+00:00',
+        '_timestamp': '2013-07-01T00:00:00+00:00',
+        'end_at': '2014-07-01T00:00:00+00:00',
+        'specific_data': 1
+      },
+      {
+        '_quarter_start_at': '2013-04-01T00:00:00+00:00',
+        '_timestamp': '2013-07-01T00:00:00+00:00',
+        'end_at': '2014-07-01T00:00:00+00:00',
+        'specific_data': 2
+      },
+      {
+        '_quarter_start_at': '2013-01-01T00:00:00+00:00',
+        '_timestamp': '2013-07-01T00:00:00+00:00',
+        'end_at': '2014-07-01T00:00:00+00:00',
+        'specific_data': 1
+      }
+    ];
+  });
+
+  describe('init', function () {
+    it('sets the dashboard slug and query options', function () {
+      var dashboard = new Dashboard('test-slug');
+
+      dashboard.options.should.eql({
+        json: true,
+        slug: 'test-slug'
+      });
+    });
+  });
+
+  describe('getConfig()', function () {
+
+    it('Should respond with a dashboard config', function () {
+
+      var testSlug = 'test-dashboard-slug';
+      var dashboard = new Dashboard(testSlug);
+
+      deferred.resolve(dashboardResponse);
+
+      return dashboard.getConfig()
+        .then(function (dashboardConfig) {
+
+          stub.should.be.calledOnce;
+          stub.getCall(0).args[0]
+            .should.eql({
+              url:
+                'https://stagecraft.production.performance.service.gov.uk/public/dashboards?slug=' +
+              testSlug,
+              json: true,
+              slug: 'test-dashboard-slug'
+            });
+
+          dashboardConfig.should.equal(dashboardResponse);
+        });
+    });
+
+  });
+
+  describe('getModule()', function () {
+
+    it('should respond with a resolved modules data', function () {
+      var dashboard = new Dashboard('test-dashboard');
+
+      deferred.resolve({
+        data: moduleDataResponse
+      });
+
+      return dashboard.getModule(module)
+        .then(function (moduleData) {
+          moduleData.data.should.equal(moduleDataResponse);
+        });
+    });
+
+  });
+
+});
